fix signature type hints.
Not sure on how to do advance classes or if it is even worth it (in python), so for e.g. S3 Client, types.Any was used

Also, for extract_findings, the function will accept either a List<dict> or dict object and wrap it accordingly. For consistency, i've added the expected list call, but this may also be called with a dict param.